requirements.txt
rasterio>=1.3
shapely>=2.0
geopandas>=0.14
pyproj>=3.6
numpy>=1.26
pandas>=2.1
scipy>=1.11
matplotlib>=3.8
tqdm>=4.66
requests>=2.31
python-dotenv>=1.0
